# services/reference-service/src/main/resources/openapi/reference-api.yaml
...
openapi: 3.1.0
info:
  title: CardDemo Reference Service API
  version: 1.0.0
  summary: >-
    Reference data of the migrated CardDemo credit-card management application:
    transaction types and their categories, disclosure-group interest rates, the
    three United States address lookups, and date evaluation.
  description: >-
    The reference context owns the data every other context reads and none of
    them writes. It publishes nineteen operations, across thirteen paths, over
    six tables.


    Transaction types and transaction categories are maintained here, and the
    relationship between them is the one referential rule the baseline states
    anywhere in this domain: a category references its type, and the type cannot
    be deleted while a category still references it. That refusal is a
    first-class user outcome in the baseline rather than a fault, and it is
    reported here as 409 for the reason recorded on the Conflict response.


    Disclosure-group rows carry the interest rate the nightly interest
    calculation consumes. That rate is an arithmetic operand, not a display
    value, so it is transported as a JSON string for the reason recorded on the
    InterestRate schema, and the lookup operation documents the DEFAULT-group
    fallback the baseline performs when an account's own group has no matching
    row.


    The three address lookups -- phone area codes, state codes and state and
...
    operation that writes them.


...


...
    one field.
...

...
servers:
...
security:
  - bearerAuth: []

# WHY : Refactoring Rationale: the authority each operation requires is declared
#       as a machine-readable field on the operation itself, and not only in
#       prose and tag membership. Prose and tag membership are not checkable: a
#       reviewer can read that a write is administrative, and a generator or a
#       gateway cannot, so the restriction would be indistinguishable from any
#       other authenticated route. Every operation below carries
#       x-required-authority.
# WHY : Assumptions: the value on each operation is not chosen here. It is
#       derived from what SecurityConfig.java already enforces for this service:
#       POST, PUT, PATCH and DELETE on every path each require
#       JwtRoleConverter.ADMIN_AUTHORITY, at its lines 99 to 106, and every
#       remaining request needs only authentication, at line 107. So every write
#       below carries carddemo-admin and every READ carries the third value
#       authenticated, and the document describes the filter chain rather than a
#       second opinion about it.
# WHY : Refactoring Rationale: an earlier revision carried carddemo-user on the
#       reads, which stated something the filter chain does not enforce and could
#       not be enforced without a behavioural regression. The shared converter
#       grants exactly the groups a token presents and builds NO hierarchy -- its
#       conversion loop admits a group only if it is one of the two recognised
#       names -- so a token carrying carddemo-admin alone receives ADMIN_AUTHORITY
#       and nothing else. A gateway or generator enforcing carddemo-user on the
#       reads would therefore refuse every administrator, and the baseline refuses
#       an administrator nothing: app/cbl/COSGN00C.cbl sends type 'A' to the
#       administrative menu at its line 245 and every function is reachable from
#       there. The honest declaration is the one the chain makes.
# WHY : Alternatives Considered: granting USER_AUTHORITY to an administrator as
#       well, so that carddemo-user could be declared and enforced on the reads.
#       Rejected because the grant would have to be made in the shared converter,
#       where it would change the authority set every one of the nine contexts
#       sees, to buy nothing this service needs: a read that any authenticated
#       caller may perform is described exactly by saying so. The two GROUP values
#       remain exactly the two the shared converter recognises, ADMIN_AUTHORITY
#       and USER_AUTHORITY at
#       services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java
#       lines 250 and 260, which are the migrated form of the two SEC-USR-TYPE
#       values 'A' and 'U'; authenticated names a REQUIREMENT rather than a group,
#       and the model below distinguishes the two.
...
paths:
...
  /api/v1/reference/disclosure-groups/{acctGroupId}/{tranTypeCd}/{tranCatCd}:

    description: >-
      One disclosure-group interest rate, addressed by the three-part key that
      identifies it: the account group, the transaction type and the transaction
      category. All three segments are required because all three form the
      primary key of reference.disclosure_groups, which
      db/migration/V1__reference.sql declares in that order.

    parameters:
      - $ref: '#/components/parameters/AcctGroupIdPath'
      - $ref: '#/components/parameters/TranTypeCdPath'
      - $ref: '#/components/parameters/TranCatCdPath'

    get:
      tags:
        - reference-data
      operationId: getDisclosureGroupRate
      x-required-authority: authenticated
      summary: Look up an interest rate, falling back to the DEFAULT group
      description: >-
        Returns the interest rate that applies to the supplied combination of
        account group, transaction type and transaction category.


        The lookup has two stages, and the second is part of the contract rather
        than an implementation detail. When no row exists for the account group
        supplied, the lookup is retried against the account group DEFAULT and the
        rate found there is returned. This reproduces
        app/cbl/CBACT04C.cbl, which on a lookup miss -- file status 23 at its
        line 436 -- moves the literal 'DEFAULT' into the group-id field at line
        437 and re-reads through the paragraph at line 443. The response states
        which of the two stages answered, in its appliedAcctGroupId and
        defaultGroupApplied members, so a caller is never left inferring whether
        the rate it received was the one it asked for.


        The fallback rows are seeded and are not optional. There are seventeen of
        them, inserted by db/migration/V2__seed_reference.sql, and they are
        carried across from app/data/ASCII/discgrp.txt, in which exactly
        seventeen of the fifty-one rows begin with the account group 'DEFAULT'
        followed by three spaces to fill the ten-character column. When the
        fallback itself finds nothing the baseline does not continue: it reports
        the condition at line 455 and abends through the paragraph invoked at
        line 458. The migrated form of that outcome is 404 carrying the abend
        detail described on the AbendDetail schema, because an absent DEFAULT row
        is a defect in the seeded reference data rather than a request a caller
        can correct.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: >-
            The applicable rate, together with the account group that actually
            supplied it. A caller that asked for a group with no row of its own
            receives the DEFAULT group's rate here with defaultGroupApplied
            true, which is a success and not a partial answer.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisclosureGroupRate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: >-
            Neither the account group supplied nor the DEFAULT group has a row
            for this transaction type and category. This is the condition the
            baseline treats as unrecoverable at app/cbl/CBACT04C.cbl lines 455
            and 458, so the body carries abend detail alongside the ordinary
            error members. The detail is the REDUCED external form: the condition
            code survives while the culprit and the reason are the empty string and
            the message is one fixed client-safe sentence, so the response names the
            seeded reference data as the subject without naming the internal
            component or its diagnostic.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/reference/date-evaluations:

    description: >-
      Evaluation of a date against a mask. This is a computation over the values
      supplied in the request, not a resource that is stored, read or listed --
      nothing is persisted and no table is consulted -- which is why the
      collection carries a single read operation and no write.

    # WHY : Assumptions: this is a GET carrying its inputs as query parameters,
    #       and NOT a POST carrying them in a body, even though a POST would read
    #       more naturally for something described as an evaluation. The reason
    #       is that the authority tier is settled by HTTP method for this whole
    #       service: SecurityConfig.java requires the administrative authority
    #       for POST, PUT, PATCH and DELETE on every path, and admits a caller
    #       holding EITHER group through its catch-all rule, which requires one
    #       of the two group authorities rather than mere authentication.
    #       Refactoring Rationale: this note previously cited
    #       anyRequest().authenticated() at a fixed line number. Both halves were
    #       liabilities -- the rule was the missing-authorization defect this
    #       contract's authority model was written against, and a line number in
    #       prose goes stale on the next edit -- so the rule is now named by what
    #       it requires and no line is quoted. Publishing this as a POST
    #       would therefore have put a side-effect-free date check behind the
    #       administrative group, so the very users who edit dated fields on the
    #       account and card screens -- the callers this evaluation exists for --
    #       would each have received 403. A GET lands on the authenticated tier,
    #       which is where the operation belongs.
    # WHY : Trade-offs: the two inputs travel in a URL, so they appear in access
    #       logs and in browser history, and the mask's literal separators must be
    #       percent-encoded by the client. That is accepted because a date and a
    #       date mask carry no cardholder-identifying value -- unlike the card
    #       numbers and account identifiers that the sibling contracts keep out of
    #       URLs deliberately -- and because being cacheable and safe to retry
    #       matches what this operation actually is.
    get:
      tags:
        - date-utilities
      operationId: evaluateDate
      x-required-authority: authenticated
      summary: Evaluate a date against a mask and return the structured verdict
      description: >-
        Evaluates the supplied date text against the supplied mask and returns
        the verdict as a structured result: a severity, a message number, the
        verdict text, and the submitted date and mask echoed back.


        This is the migrated form of app/cbl/CSUTLDTC.cbl, which takes exactly
        these inputs and produces exactly this result. Its linkage declares the
        date and the mask as ten-character fields and the result as eighty
        characters, at lines 84 to 86, and receives all three at line 88; it
        calls the platform date service through the five-line invocation at lines
        116 to 120 and maps that service's feedback onto the outcome set
        described on the DateFeedbackCode schema. The five components of this
        response are the five variable fields of that eighty-character result,
        interleaved there with the constant labels declared at its lines 45, 51
        and 54.


        The rules themselves are not restated here and are not implemented by
        this service. They live in the shared kernel, at
        services/common-lib/src/main/java/com/carddemo/common/validation/DateEditValidator.java,
        whose entry point takes a date and a mask and returns this result. This
        contract describes the request shape, the accepted masks and the result
        shape, and nothing beyond that.


        A mask this service does not recognise is not a request error: it is
        answered with 200 carrying the verdict for an unusable mask pattern --
        feedback code BAD_PIC_STRING, severity 3, message number 2518 -- because the
        baseline date service declares a condition for exactly that case at
        app/cbl/CSUTLDTC.cbl line 68 and returns it like any other feedback. That is
        why the mask parameter is typed as DateMaskInput, a bounded ten-character
        string with no enumeration: an enumerated input would have made this outcome
        unreachable, since a request validator refuses a value outside an enumeration
        before any evaluation runs. The masks this service DOES recognise are
        published separately, on the DateMask schema, so that a caller can send one
        it knows will be interpreted. A date whose width does not match the
        recognised mask it was sent with IS a request error, and is answered with
        400, because the components are read by offset and a shorter text has no
        components to read.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - $ref: '#/components/parameters/DateValueQuery'
        - $ref: '#/components/parameters/DateMaskQuery'
      responses:
        '200':
          description: >-
            The verdict. A severity of zero means the date was accepted; any
            other severity means it was rejected and the verdict text names the
            reason. Note the trap recorded on the DateFeedbackCode schema: the
            accepting outcome's own condition name in the baseline reads as
            though it meant the opposite.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DateEvaluationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A JSON Web Token issued by the configured identity provider and validated
        as a resource-server token. Its cognito:groups claim is converted to
        authorities by
        services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java,
        where the baseline user type 'A' becomes carddemo-admin and 'U' becomes
        carddemo-user. Presenting no usable token yields 401; presenting a valid
        token that lacks the authority an operation requires yields 403.
  parameters:
...
    AcctGroupIdPath:
      name: acctGroupId
      in: path
      required: true
      description: >-
        The account group whose rate is sought. Send it without trailing padding;
        the stored column is ten characters and pads on comparison, so 'DEFAULT'
        and its padded form address the same rows.
      schema:
        $ref: '#/components/schemas/AccountGroupId'
    TranTypeCdPath:
      name: tranTypeCd
      in: path
      required: true
      description: >-
        The two-digit transaction type code of the rate being sought. It is the
        second component of the disclosure-group key.
      schema:
        $ref: '#/components/schemas/TransactionTypeCode'
    TranCatCdPath:
      name: tranCatCd
      in: path
      required: true
      description: >-
        The four-digit transaction category code of the rate being sought. It is
        the third component of the disclosure-group key.
      schema:
        $ref: '#/components/schemas/TransactionCategoryCode'
...
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >-
        Identity of the unit of work this request belongs to, propagated into the
        service log record and echoed on the response. Supply a value to tie this
        request to work already in progress upstream; omit it and the service
        mints one. A value that does not fit the published bound and character set
        is refused with 400 before any handler runs. A value whose shape is that of a
        primary account number -- only digits and the three separators, with
        thirteen or more digits -- is refused on the same terms, because this
        identity is echoed and logged. The refusal carries the shared problem body
        and the response correlation header, the same as any other 400, and it names
        the rule it applied without ever repeating the value it refused.
      schema:
        $ref: '#/components/schemas/CorrelationId'
  headers:
    XCorrelationId:
      description: >-
        Identity of the unit of work that produced this response. Echoed from the
        request where the caller supplied a conforming value, and minted by the
        service where it did not. Quote it when reporting a failure, because it is
        the route from this response to the operational record.
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'

    # WHY : Refactoring Rationale: this header is declared because the service
    #       actually sends it and an earlier revision of this document did not say
    #       so. The 401 is rendered by the shared handler
    #       common/error/ApiErrorSecurityHandlers.UnauthenticatedEntryPoint, which
    #       delegates to the framework's BearerTokenAuthenticationEntryPoint FIRST
    #       and only then writes the body, precisely so this challenge is composed
    #       exactly as the OAuth 2.0 bearer-token specification requires. A client's
    #       token-refresh logic reads this header, so leaving it undeclared left a
    #       generated client with no typed access to the one header it needs on the
    #       one response it needs it on.
    # WHY : Assumptions: the value is not enumerated and no pattern is written for
    #       it. The framework composes it from the error it detected -- a bare
    #       scheme when no token was presented, and a scheme with an error code and
    #       description when one was presented and rejected -- so pinning a shape
    #       here would either be wrong for one of those two cases or would have to
    #       be widened until it constrained nothing.
...
  responses:
    BadRequest:
      description: >-
        The request could not be accepted as supplied. This covers a malformed or
        out-of-range path or query value, a body that fails a declared constraint,
        a cursor that cannot be opened or was replayed against the wrong filter,
        collection or direction, a correlation header outside the published bound,
        and a write whose submitted values are identical to the ones already stored.
        The fieldErrors array names every offending field rather than only the first,
        for the reason recorded on the FieldError schema.


        The unchanged-submission refusal is the one member of that list that is a
        business rule rather than a syntactic check, and it belongs here rather than
        on 409 because it is decided from the request alone, before any row is
        contended: the baseline compares the submitted values with the ones it
        fetched and re-displays without writing. Its sentence is carried verbatim and
        is published on the ReferenceMessageCatalogue schema in both of the two
        wordings the baseline declares -- noChangeDetectedSingleRecord for a
        single-record replace and noChangeDetectedList for the list screen -- with a
        field entry naming the member that was resubmitted unchanged.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    # WHY : Refactoring Rationale: this response now states WHICH code and sentence
    #       the body carries, because an earlier revision promised the shared problem
    #       shape without naming its contents and the security chain did not produce
    #       one at all. A refusal decided by the filter chain never reaches a
    #       controller, so it never reaches the shared @RestControllerAdvice, and the
    #       framework default answered with a status and a challenge header and no
    #       body -- the exact opposite of what this response declared. The chain now
    #       installs common/error/ApiErrorSecurityHandlers, whose 401 renderer emits
    #       CARDDEMO-0401 with the sentence below.
    # WHY : Assumptions: the correlation identifier in the body is the SAME value as
    #       the header beside it and as the log line for this refusal, the renderer
    #       reading it from the diagnostic context the correlation filter established
    #       ahead of the security chain. That equality is the whole reason the header
    #       is required on a response whose body names no reason.
    Unauthorized:
      description: >-
        No usable bearer token was presented: the header was absent, the token was
        malformed, its signature or issuer did not validate, or it had expired.
        This is distinct from 403 and from 409, and the distinction is not
        cosmetic -- a caller acts on 401 by obtaining a token, on 403 by obtaining
        a different one, and on 409 by changing its request. No body member names
        the group the caller lacks, because at this point no verified identity
        exists to compare against one. The body carries code CARDDEMO-0401 and the
        sentence "Sign on to continue", identically on all four of those causes,
        alongside the bearer-token challenge below.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
        WWW-Authenticate:
          $ref: '#/components/headers/WwwAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: >-
        A valid token was presented, but it does not carry the authority this
        operation requires. In practice this is a token holding only carddemo-user
        sent to one of the administrative write operations, which
        SecurityConfig.java restricts by method at its lines 99 to 106. The
        identity is established and the request is well formed; only the authority
        is missing, which is why this is neither 401 nor 400. The body carries code
        CARDDEMO-0403 and the sentence "Not authorized for this operation" -- the
        same pair the shared advice publishes for a denial decided by a
        method-security check, so a caller cannot tell from the response where in
        the stack the authority was checked.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    DisclosureGroupRate:
      type: object
      description: >-
        The interest rate that applies to a combination of account group,
        transaction type and category, together with a statement of which account
        group actually supplied it. This is the migrated form of the fifty-byte
        record app/cpy/CVTRA02Y.cpy declares, less the twenty-eight-byte trailing
        filler at its line 10 -- the one filler of that record -- with two members
        added that the baseline has no room for, because the baseline conveys the
        same information by having overwritten its own key field before re-reading.
      required:
        - requestedAcctGroupId
        - appliedAcctGroupId
        - tranTypeCd
        - tranCatCd
        - interestRate
        - defaultGroupApplied
      additionalProperties: false
      properties:
        requestedAcctGroupId:
          allOf:
            - $ref: '#/components/schemas/AccountGroupId'
          description: >-
            The account group the caller asked about, echoed back unchanged.
        # WHY : Refactoring Rationale: the group that answered is reported
        #       EXPLICITLY, where the baseline leaves it implicit. On a miss
        #       app/cbl/CBACT04C.cbl overwrites its own group-id field with the
        #       literal 'DEFAULT' at line 437 before re-reading, so after the
        #       second read the field no longer records which group was asked for
        #       and nothing downstream can tell a fallback rate from a specific
        #       one. Returning both values keeps the same rate selection while
        #       making the substitution visible, so a caller reconciling a charge
        #       against a rate table is not left inferring it. The divergence is
        #       registered as D-APPLIED-GROUP-VISIBLE in
        #       docs/architecture/cobol-to-service-traceability.md.
        appliedAcctGroupId:
          allOf:
            - $ref: '#/components/schemas/AccountGroupId'
          description: >-
            The account group whose row supplied the rate returned. Equal to
            requestedAcctGroupId when that group had a row of its own, and
            'DEFAULT' when the fallback answered.
        tranTypeCd:
          allOf:
            - $ref: '#/components/schemas/TransactionTypeCode'
          description: >-
            The transaction type of the rate returned, echoed back unchanged. The
            fallback substitutes only the account group, never this.
        tranCatCd:
          allOf:
            - $ref: '#/components/schemas/TransactionCategoryCode'
          description: >-
            The transaction category of the rate returned, echoed back unchanged.
        interestRate:
          allOf:
            - $ref: '#/components/schemas/InterestRate'
          description: >-
            The annual rate that applies, as an exact decimal string. It is the
            operand of the interest computation, so parse it into a decimal type.
        defaultGroupApplied:
          type: boolean
          description: >-
            Whether the fallback answered. True exactly when appliedAcctGroupId is
            'DEFAULT' and the caller asked about a different group; false when the
            requested group had its own row, including when the caller asked about
            'DEFAULT' directly.
      examples:
        - requestedAcctGroupId: A000000000
          appliedAcctGroupId: A000000000
          tranTypeCd: '01'
          tranCatCd: '0001'
          interestRate: '15.00'
          defaultGroupApplied: false
        - requestedAcctGroupId: Z999999999
          appliedAcctGroupId: DEFAULT
          tranTypeCd: '01'
          tranCatCd: '0002'
          interestRate: '25.00'
          defaultGroupApplied: true
...